gemini-chat-langchain/requirements.txt
langchain
langchain-google-vertexai
python-dotenv
urllib3==1.26.18